requirements.txt
attrs
cachetools
certifi
chardet
click
fastapi
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-resumable-media
googleapis-common-protos
h11
httptools
idna
importlib-metadata
more-itertools
numpy
packaging
pandas
pluggy
protobuf
py
pyasn1
pyasn1-modules
pydantic
pyparsing
pytest
python-dateutil
pytz
requests
rsa
ruamel.yaml
ruamel.yaml.clib
six
SQLAlchemy
starlette
ujson
urllib3
uvicorn
uvloop
wcwidth
websockets
zipp
